Address PR review critiques
* Factor out Test.Goblin.Persist module.
* Drop unnecessary function.
* Rename mk*Goblin functions for consistency.
